BUG NEEDS FIXING: Installed more dependencies and tried to render an exams table.
I tried to render a table with a particular patient's exams. I don't have all the columns just yet, but I just wanted to see if I can render a table. However, there is a bug in line 26 of client/pages/patientDetails that is not allowing me to call the API function getExamsOfPatient.